modified finalname of tar.gz packages
new apache style name was 'apache-druid-<version>', but this messes with
our current chef-cookbook for installing and configuring druid. So we
want to keep the name as 'druid-<verision>'.

So to ease testing, this needs to be in here for the time being. When
both production and accept environment are updated, we can resort to
having the default name again.